Qualities of a good leader
Love and compassion
The most essential qualities of a good leader are love and compassion. Love and compassion are inseparable qualities of a good leader. It is love and compassion that would motivate the good leader to do those things that would boost the standard of a living of a citizens.
The reason some leaders pay deaf ear to the plights of their subjects is simply lack of love and compassion for them (the masses or subjects). And this is a direct opposite of the leadership style of Jesus (the greatest leader) whose leadership style was full of love and compassion.
The Jews testify of Jesus' love toward Lazarus. Jesus feeds five thousand men with seven loaves of bread and few fishes out of love and compassion.